Island: Move app authentication setup in authentication_service
