+++ monkey/monkey_island/cc/services/__init__.py
@@ -14,3 +14,4 @@ from .authentication_service import (
     register_resources as register_authentication_resources,
 )  # noqa: E501
 from .authentication_service import build as build_authentication_service
+from .authentication_service import setup_authentication
