+++ monkey/monkey_island/cc/services/authentication_service/__init__.py
@@ -2,3 +2,4 @@ from .authentication_service import AuthenticationService
 from .flask_resources import register_resources
 
 from .build import build
+from .configure_flask_security import setup_authentication
